Update streamlit_app.py
put the list of selected fruits into a variable called fruits_selected. Then, use fruits in fruits_selected to pull rows from full data set (and assign that data to a variable called fruits_to_show). Finally, use the data in fruits_to_show in the dataframe it displays on the page.